+++ run.sh
@@ -1,0 +1,3 @@
+sudo docker start mongodb
+sudo docker start myredis
+npm start