+++ src/modules/meal/meal.types.ts
@@ -1,4 +1,5 @@
 import { IngredientType, MealIngredient } from '../ingredient/ingredient.types';
+import { TranslatedIngredient } from '../translation/translation.types';
 
 export type RatedMeal = {
     id: string,
@@ -8,12 +9,14 @@ export type RatedMeal = {
     relevance: number
 };
 
-type MealRecipeSection = {
+export type MealRecipeStep = {
+    number: number,
+    step: string
+};
+
+export type MealRecipeSection = {
     name?: string,
-    steps: {
-        number: number,
-        step: string
-    }[]
+    steps: MealRecipeStep[]
 };
 
 export type MealRecipeSections = MealRecipeSection[];
@@ -35,6 +38,12 @@ export type DetailedMeal = {
     },
     recipeSections: MealRecipeSections
 };
+
+export interface DetailedMealWithTranslations {
+    meal: DetailedMeal;
+    ingredients: TranslatedIngredient[];
+    recipe: MealRecipeSections;
+}
 
 export type ProposedMeal = {
     id: string,